Rewrite syncing
Simplify the locking by only reading/writing when we the lock counter is at 0,
otherwise keep a local count.

With the merge state we will be a little more primitive and write every time
there is an update. The state can change even if we have the lock, other merges
will still be progressing, and it's simpler not to attempt to cache that in
memory. Yes, file read/write takes a long time compared to in-memory ops, but
totally insignificant compared to waiting the bottlenecks in this process, i.e.
waiting for CI to pass.

This requires explicitly pulling from remote so that method is exposed, but it's
not tenable to attempt to do this automatically within the GitRepo class.