Remove singletons
We were using a singleton pattern to ensure that objects such as Lock and
QueueState maintained their state. But it was becoming problematic with regards
to circular dependencies and testing.

Instead we create a single instance in the main MergeQueue class and then pass
a reference to that into all other objects, so they can use the main class as an
accessor to othe other objects they need.