Simplify config
Get rid of the Configurable module. It was always a bit confusing, and doing two
different things. Just create an instance of Config in the main app class and
delegate to that.

The Config class now has defaults, which can be overriden by ENV var, which can
finally be overriden by params passed into its initializer.