Repoint pr branch to merge branch before merge
Previously we were rebasing the PR branch onto main, however this would create a
new sha which hadn't yet passed CI, and would therefore cause the PR to fail the
CI check.

Instead now we hard reset the PR branch to the merge branch. The resulting code
is the same, but now the sha has already passed CI and can be merged
straightaway.